Rename TableN to Table
#2559
