+++ src/settings/GlobalSettingsViewer.ts
@@ -28,8 +28,8 @@ import {Icons} from "../gui/base/icons/Icons"
 import {showProgressDialog} from "../gui/dialogs/ProgressDialog"
 import type {EntityUpdateData} from "../api/main/EventController"
 import {isUpdateForTypeRef} from "../api/main/EventController"
-import type {TableAttrs, TableLineAttrs} from "../gui/base/TableN"
-import {ColumnWidth, createRowActions} from "../gui/base/TableN"
+import type {TableAttrs, TableLineAttrs} from "../gui/base/Table.js"
+import {ColumnWidth, createRowActions} from "../gui/base/Table.js"
 import {attachDropdown, createDropdown, DropdownChildAttrs} from "../gui/base/Dropdown.js"
 import {ButtonType} from "../gui/base/Button.js"
 import {DomainDnsStatus} from "./DomainDnsStatus"
